requirements.txt
aiohttp==3.5.4
astroid==2.3.3
async-timeout==3.0.1
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
darksky-weather==1.8.0
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
multidict==4.7.5
pylint==2.4.4
pytz==2019.1
requests==2.21.0
six==1.14.0
typed-ast==1.4.1
urllib3==1.24.3
wrapt==1.11.2
yarl==1.4.2
